+++ src/main/java/br/com/fiap/restaurantes/gerenciamento/infra/repository/UsuarioRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UsuarioRepository extends JpaRepository<UsuarioEntity, Long> {
+public interface UsuarioRepository extends JpaRepository<UsuarioEntity, Integer> {
 
     boolean existsByLogin(String login);
     Optional<UsuarioEntity> findByLogin(String login);
